Rename main package
