+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.bestdamn.fortyk.crusade"
+        applicationId "com.stoffer.fortyk.crusade"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
