+++ app/src/androidTest/java/com/stoffer/fortyk/crusade/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.bestdamn.fortyk.crusade
+package com.stoffer.fortyk.crusade
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.bestdamn.fortyk.crusade", appContext.packageName)
+        assertEquals("com.stoffer.fortyk.crusade", appContext.packageName)
     }
 }
