+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.bestdamn.fortyk.crusade">
+    package="com.stoffer.fortyk.crusade">
 
     <application
         android:allowBackup="true"
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".UnitActivity"/>
-        <activity android:name=".ForceAcitivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name="com.stoffer.fortyk.crusade.UnitActivity"/>
+        <activity android:name="com.stoffer.fortyk.crusade.ForceAcitivity" />
+        <activity android:name="com.stoffer.fortyk.crusade.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
