+++ app/src/main/java/com/stoffer/fortyk/crusade/ForceAcitivity.kt
@@ -1,4 +1,4 @@
-package com.bestdamn.fortyk.crusade
+package com.stoffer.fortyk.crusade
 
 import android.content.Intent
 import android.widget.Toast
@@ -7,9 +7,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.bestdamn.fortyk.crusade.databinding.ActivityForceBinding
-import com.bestdamn.fortyk.crusade.domain.Force
-import com.bestdamn.fortyk.crusade.domain.Unit
+import com.stoffer.fortyk.crusade.databinding.ActivityForceBinding
+import com.stoffer.fortyk.crusade.domain.Force
+import com.stoffer.fortyk.crusade.domain.Unit
 import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.AdView
 import com.google.gson.Gson
