+++ app/src/main/java/com/stoffer/fortyk/crusade/ForceAdapter.kt
@@ -1,7 +1,7 @@
-package com.bestdamn.fortyk.crusade
+package com.stoffer.fortyk.crusade
 
 import android.app.AlertDialog
-import android.content.Context
+import android.content.Context.MODE_PRIVATE
 import android.content.DialogInterface
 import android.content.Intent
 import android.view.LayoutInflater
@@ -9,29 +9,28 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import androidx.recyclerview.widget.RecyclerView
-import com.bestdamn.fortyk.crusade.domain.Force
-import com.bestdamn.fortyk.crusade.domain.Unit
+import com.stoffer.fortyk.crusade.domain.Force
 import com.google.gson.Gson
-import kotlinx.android.synthetic.main.unit_list_text_view.view.*
+import kotlinx.android.synthetic.main.force_list_text_view.view.*
 
-class UnitAdapter(private var myDataset: Array<Unit>) :
-    RecyclerView.Adapter<UnitAdapter.MyViewHolder>() {
+class ForceAdapter(private var myDataset: Array<Force>) :
+    RecyclerView.Adapter<ForceAdapter.MyViewHolder>() {
 
     // Provide a reference to the views for each data item
     // Complex data items may need more than one view per item, and
     // you provide access to all the views for a data item in a view holder.
     // Each data item is just a string in this case that is shown in a TextView.
-    class MyViewHolder(val unitLayout: LinearLayout) : RecyclerView.ViewHolder(unitLayout)
+    class MyViewHolder(val forceLayout: LinearLayout) : RecyclerView.ViewHolder(forceLayout)
 
 
     // Create new views (invoked by the layout manager)
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int
-    ): UnitAdapter.MyViewHolder {
+    ): ForceAdapter.MyViewHolder {
         // create a new view
         val textView = LayoutInflater.from(parent.context)
-            .inflate(R.layout.unit_list_text_view, parent, false) as LinearLayout
+            .inflate(R.layout.force_list_text_view, parent, false) as LinearLayout
         return MyViewHolder(textView)
     }
 
@@ -39,36 +38,37 @@ class UnitAdapter(private var myDataset: Array<Unit>) :
     override fun onBindViewHolder(holder: MyViewHolder, position: Int) {
         // - get element from your dataset at this position
         // - replace the contents of the view with that element
-        holder.unitLayout.textViewUnitName.text = myDataset[position].name
-        holder.unitLayout.textViewUnitType.text = myDataset[position].type
+        holder.forceLayout.textViewForceName.text = myDataset[position].name
+        holder.forceLayout.textViewForceFaction.text = myDataset[position].faction
 
         holder.itemView.setOnClickListener(View.OnClickListener {
             val gson = Gson()
             val context = it.context
-            val unitJson = gson.toJson(myDataset[position])
-            val unitIntent = Intent(context, UnitActivity::class.java)
-            unitIntent.putExtra("unit", unitJson)
-            context.startActivity(unitIntent)
+            val forceJson = gson.toJson(myDataset[position])
+            val forceIntent = Intent(context, ForceAcitivity::class.java)
+            forceIntent.putExtra("force", forceJson)
+            context.startActivity(forceIntent)
         })
 
         holder.itemView.setOnLongClickListener(
             View.OnLongClickListener {
                 val builder = AlertDialog.Builder(it.context).apply {
-                    setTitle("Delete Unit?")
-                    setMessage("Are you sure you want to delete the Unit: " + myDataset[position].name + "?")
+                    setTitle("Delete Force?")
+                    setMessage("Are you sure you want to delete the Force: " + myDataset[position].name + "?")
                     setPositiveButton("Yes", DialogInterface.OnClickListener { dialog, which ->
-                        val sharedPrefs = it.context.getSharedPreferences(
-                            "sharedPrefs",
-                            Context.MODE_PRIVATE
-                        )
+
+                        val sharedPrefs =
+                            it.context.getSharedPreferences("sharedPrefs", MODE_PRIVATE)
                         val editor = sharedPrefs.edit()
-                        val forceJson = sharedPrefs.getString(myDataset[position].force_id, null)
-                        val gson = Gson()
-                        val force = gson.fromJson(forceJson, Force::class.java)
-                        val unitToRemove = force.units.find { unit -> unit.id == myDataset[position].id }
-                        force.units.remove(unitToRemove)
-                        val updatedForceJson = gson.toJson(force)
-                        editor.putString(force.id, updatedForceJson)
+
+                        // remove force id from force ids list
+                        val forceSet = sharedPrefs.getStringSet("forceIds", null)
+                        forceSet?.remove(myDataset[position].id)
+                        editor.putStringSet("forceIds", forceSet)
+
+                        // remove the force itself
+                        editor.remove(myDataset[position].id)
+
                         editor.apply()
 
                         val list = myDataset.toMutableList()
@@ -76,6 +76,7 @@ class UnitAdapter(private var myDataset: Array<Unit>) :
                         myDataset = list.toTypedArray()
                         notifyItemChanged(position)
                         notifyItemRemoved(position)
+
                     })
                     setNegativeButton("No", null)
