+++ app/src/main/java/com/stoffer/fortyk/crusade/UnitAdapter.kt
@@ -1,7 +1,7 @@
-package com.bestdamn.fortyk.crusade
+package com.stoffer.fortyk.crusade
 
 import android.app.AlertDialog
-import android.content.Context.MODE_PRIVATE
+import android.content.Context
 import android.content.DialogInterface
 import android.content.Intent
 import android.view.LayoutInflater
@@ -9,28 +9,29 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.LinearLayout
 import androidx.recyclerview.widget.RecyclerView
-import com.bestdamn.fortyk.crusade.domain.Force
+import com.stoffer.fortyk.crusade.domain.Force
+import com.stoffer.fortyk.crusade.domain.Unit
 import com.google.gson.Gson
-import kotlinx.android.synthetic.main.force_list_text_view.view.*
+import kotlinx.android.synthetic.main.unit_list_text_view.view.*
 
-class ForceAdapter(private var myDataset: Array<Force>) :
-    RecyclerView.Adapter<ForceAdapter.MyViewHolder>() {
+class UnitAdapter(private var myDataset: Array<Unit>) :
+    RecyclerView.Adapter<UnitAdapter.MyViewHolder>() {
 
     // Provide a reference to the views for each data item
     // Complex data items may need more than one view per item, and
     // you provide access to all the views for a data item in a view holder.
     // Each data item is just a string in this case that is shown in a TextView.
-    class MyViewHolder(val forceLayout: LinearLayout) : RecyclerView.ViewHolder(forceLayout)
+    class MyViewHolder(val unitLayout: LinearLayout) : RecyclerView.ViewHolder(unitLayout)
 
 
     // Create new views (invoked by the layout manager)
     override fun onCreateViewHolder(
         parent: ViewGroup,
         viewType: Int
-    ): ForceAdapter.MyViewHolder {
+    ): UnitAdapter.MyViewHolder {
         // create a new view
         val textView = LayoutInflater.from(parent.context)
-            .inflate(R.layout.force_list_text_view, parent, false) as LinearLayout
+            .inflate(R.layout.unit_list_text_view, parent, false) as LinearLayout
         return MyViewHolder(textView)
     }
 
@@ -38,37 +39,36 @@ class ForceAdapter(private var myDataset: Array<Force>) :
     override fun onBindViewHolder(holder: MyViewHolder, position: Int) {
         // - get element from your dataset at this position
         // - replace the contents of the view with that element
-        holder.forceLayout.textViewForceName.text = myDataset[position].name
-        holder.forceLayout.textViewForceFaction.text = myDataset[position].faction
+        holder.unitLayout.textViewUnitName.text = myDataset[position].name
+        holder.unitLayout.textViewUnitType.text = myDataset[position].type
 
         holder.itemView.setOnClickListener(View.OnClickListener {
             val gson = Gson()
             val context = it.context
-            val forceJson = gson.toJson(myDataset[position])
-            val forceIntent = Intent(context, ForceAcitivity::class.java)
-            forceIntent.putExtra("force", forceJson)
-            context.startActivity(forceIntent)
+            val unitJson = gson.toJson(myDataset[position])
+            val unitIntent = Intent(context, UnitActivity::class.java)
+            unitIntent.putExtra("unit", unitJson)
+            context.startActivity(unitIntent)
         })
 
         holder.itemView.setOnLongClickListener(
             View.OnLongClickListener {
                 val builder = AlertDialog.Builder(it.context).apply {
-                    setTitle("Delete Force?")
-                    setMessage("Are you sure you want to delete the Force: " + myDataset[position].name + "?")
+                    setTitle("Delete Unit?")
+                    setMessage("Are you sure you want to delete the Unit: " + myDataset[position].name + "?")
                     setPositiveButton("Yes", DialogInterface.OnClickListener { dialog, which ->
-
-                        val sharedPrefs =
-                            it.context.getSharedPreferences("sharedPrefs", MODE_PRIVATE)
+                        val sharedPrefs = it.context.getSharedPreferences(
+                            "sharedPrefs",
+                            Context.MODE_PRIVATE
+                        )
                         val editor = sharedPrefs.edit()
-
-                        // remove force id from force ids list
-                        val forceSet = sharedPrefs.getStringSet("forceIds", null)
-                        forceSet?.remove(myDataset[position].id)
-                        editor.putStringSet("forceIds", forceSet)
-
-                        // remove the force itself
-                        editor.remove(myDataset[position].id)
-
+                        val forceJson = sharedPrefs.getString(myDataset[position].force_id, null)
+                        val gson = Gson()
+                        val force = gson.fromJson(forceJson, Force::class.java)
+                        val unitToRemove = force.units.find { unit -> unit.id == myDataset[position].id }
+                        force.units.remove(unitToRemove)
+                        val updatedForceJson = gson.toJson(force)
+                        editor.putString(force.id, updatedForceJson)
                         editor.apply()
 
                         val list = myDataset.toMutableList()
@@ -76,7 +76,6 @@ class ForceAdapter(private var myDataset: Array<Force>) :
                         myDataset = list.toTypedArray()
                         notifyItemChanged(position)
                         notifyItemRemoved(position)
-
                     })
                     setNegativeButton("No", null)
