+++ app/src/main/java/com/stoffer/fortyk/crusade/domain/Force.kt
@@ -1,4 +1,4 @@
-package com.bestdamn.fortyk.crusade.domain
+package com.stoffer.fortyk.crusade.domain
 
 import java.util.*
