+++ app/src/main/res/layout/activity_force.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="force"
-            type="com.bestdamn.fortyk.crusade.domain.Force" />
+            type="com.stoffer.fortyk.crusade.domain.Force" />
     </data>
 
     <LinearLayout
