+++ app/src/main/res/layout/activity_unit.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="unit"
-            type="com.bestdamn.fortyk.crusade.domain.Unit" />
+            type="com.stoffer.fortyk.crusade.domain.Unit" />
     </data>
 
     <LinearLayout
